changed int to float

+++ programming_paradigm/robust_division_calculator.py
@@ -8,7 +8,7 @@ logic including error handling, and main.py, which interfaces with the user thro
 
 def safe_divide(numerator, denominator):
     try: 
-        result = int(numerator) / int(denominator)
+        result = float(numerator) / float(denominator)
         return f"The result of the division is: {result}."
     except ZeroDivisionError:
         print("Error: Cannot divide by zero")
